create new worker
