+++ lib/worker-server.js
@@ -16,6 +16,7 @@ module.exports = new ponos.Server({
     }
   },
   tasks: {
+    'docker.container.poll': require('./workers/docker.container.poll.js'),
     'docker.event.publish': require('./workers/docker.event.publish.js'),
     'docker.events-stream.connect': require('./workers/docker.events-stream.connect.js'),
     'swarm.events-stream.connect': require('./workers/swarm.events-stream.connect.js')
