+++ lib/workers/docker.event.publish.js
@@ -7,6 +7,7 @@ const Promise = require('bluebird')
 const WorkerStopError = require('error-cat/errors/worker-stop-error')
 
 const datadog = require('../datadog')
+const dockerUtils = require('../docker-utils')
 const Docker = require('../docker')
 const Swarm = require('../swarm')
 const logger = require('../logger')
@@ -49,7 +50,7 @@ const DockerEventPublish = (job) => {
           return job
         })
         .catch((err) => {
-          return DockerEventPublish._handleInspectError(job.Host, err, log)
+          return dockerUtils._handleInspectError(job.Host, err, log)
         })
     })
     .then((event) => {
@@ -63,54 +64,6 @@ DockerEventPublish._isUserContainer = (event) => {
 }
 DockerEventPublish._isBuildContainer = (event) => {
   return keypather.get(event, 'inspectData.Config.Labels.type') === 'image-builder-container'
-}
-
-/**
- * handles inspect errors.
- * if 404 call task fatal
- *   else check if the dock still exists
- *   throw task fatal if it does not
- * @param  {String} host target host for inspect
- * @param  {Error}  err  error from docker inspect
- * @param  {Object} log  current logger
- * @return {Promise}
- * @resolve {undefined}
- * @reject {WorkerStopError} If 404 or dock does not exist
- *         {Error}          If error unknown
- * @throws {WorkerStopError} If 404 or dock does not exist
- */
-DockerEventPublish._handleInspectError = (host, err, log) => {
-  if (err.statusCode === 404) {
-    // container is not there anymore. Exit
-    const fatalErr = new WorkerStopError(
-      err.message,
-      { originalError: err }, { level: 'info' }
-    )
-    log.trace({ err: err }, 'DockerEventPublish - container not found')
-    throw fatalErr
-  }
-  log.error({ err: err }, 'DockerEventPublish - inspect error')
-  // check to see if host still exist before retrying
-  const swarm = new Swarm(process.env.SWARM_HOST)
-  return swarm.swarmHostExistsAsync(host)
-    .catch((hostErr) => {
-      log.trace({ err: hostErr }, 'DockerEventPublish - swarmHostExists error')
-      // if above errors, throw original error
-      throw err
-    })
-    .then((exists) => {
-      if (!exists) {
-        log.trace('DockerEventPublish - host does not exist')
-        const fatalErr = new WorkerStopError(
-          'host does not exist',
-          { host: host }
-        )
-        fatalErr.report = false
-        throw fatalErr
-      }
-
-      throw err
-    })
 }
 
 /**
